hoit

+++ libft/ft_atoi.c
@@ -1,39 +1,47 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_atoi.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: guntkim <guntkim@student.42seoul.k>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/03/15 15:41:59 by guntkim           #+#    #+#             */
+/*   Updated: 2022/03/15 18:14:50 by guntkim          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-int ft_isspace(const char c)
+int	ft_isspace(const char c)
 {
-    if (c >= 9 && c <= 13)
-        return (1);
-    if (c == 32)
-        return (1);
-    return (0);
+	if (c >= 9 && c <= 13)
+		return (1);
+	if (c == 32)
+		return (1);
+	return (0);
 }
 
-int ft_atoi(const char *str)
+int	ft_atoi(const char *str)
 {
-    unsigned int    num;
-    int             sign;
-    int             index;
+	long long	num;
+	int			sign;
+	int			index;
 
-    num = 0;
-    sign = 1;
-    index = 0;
-
-    // Skip White Space
-    while (str[index] && ft_isspace(str[index]) != 0)
-        index += 1;
-    
-    // Process Sign
-    if (str[index] == '+' || str[index] == '-')
-    {
-        if (str[index] == '-')
-            sign *= -1;
-        index += 1;
-    }
-    while (str[index] >= '0' && str[index] <= '9')
-    {
-        num = (num * 10) + (str[index] - '0');
-        index += 1;
-    }
-    return (num * sign);
+	num = 0;
+	sign = 1;
+	index = 0;
+	while (str[index] && ft_isspace(str[index]) != 0)
+		index += 1;
+	if (str[index] == '+' || str[index] == '-')
+	{
+		if (str[index] == '-')
+			sign *= -1;
+		index += 1;
+	}
+	while (str[index] >= '0' && str[index] <= '9')
+	{
+		num = (num * 10) + (str[index] - '0');
+		index += 1;
+	}
+	return ((int)num * sign);
 }
