+++ libft/ft_strdup.c
@@ -1,19 +1,31 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strdup.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: guntkim <guntkim@student.42seoul.k>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/03/15 15:47:07 by guntkim           #+#    #+#             */
+/*   Updated: 2022/03/15 18:52:09 by guntkim          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-char    *ft_strdup(const char *s1)
+char	*ft_strdup(const char *s1)
 {
-    int     i;
-    char    *dest;
-    
-    i = 0;
-    dest = (char *)malloc(sizeof(char) * ft_strlen(s1) + 1);
-    if (!dest)
-        return (0);
-    while (s1[i])
-    {
-        dest[i] = s1[i];
-        i += 1;
-    }
-    dest[i] = '\0';
-    return (dest);
+	size_t	i;
+	char	*dest;
+
+	i = 0;
+	dest = (char *)malloc(sizeof(char) * ft_strlen(s1) + 1);
+	if (!dest)
+		return (0);
+	while (s1[i])
+	{
+		dest[i] = s1[i];
+		i += 1;
+	}
+	dest[i] = '\0';
+	return (dest);
 }
